requirements.txt
agno==1.1.3
streamlit
google-generativeai
pydantic
duckduckgo-search
lancedb
opencv-python-headless
google-genai
youtube-transcript-api>=0.6.1
pytube>=15.0.0
streamlit>=1.32.0
python-dotenv>=1.0.0